Extract APIRequestLoggingFormatter
We're having this in the new TeacherTrainingAPI namespace
